requirements.txt
PyPDF2==3.0.1
google-api-python-client==2.108.0
google-auth==2.27.0
google-auth-oauthlib==1.1.0
streamlit==1.32.2
openai==1.17.0
python-dotenv==1.0.1